Extract exhaustiveness into its own crate
