+++ compiler/rustc_pattern_analysis/Cargo.toml
@@ -1,27 +1,22 @@
 [package]
-name = "rustc_mir_build"
+name = "rustc_pattern_analysis"
 version = "0.0.0"
 edition = "2021"
 
 [dependencies]
 # tidy-alphabetical-start
-either = "1"
 rustc_apfloat = "0.2.0"
 rustc_arena = { path = "../rustc_arena" }
-rustc_ast = { path = "../rustc_ast" }
 rustc_data_structures = { path = "../rustc_data_structures" }
 rustc_errors = { path = "../rustc_errors" }
 rustc_fluent_macro = { path = "../rustc_fluent_macro" }
 rustc_hir = { path = "../rustc_hir" }
 rustc_index = { path = "../rustc_index" }
-rustc_infer = { path = "../rustc_infer" }
 rustc_macros = { path = "../rustc_macros" }
 rustc_middle = { path = "../rustc_middle" }
-rustc_pattern_analysis = { path = "../rustc_pattern_analysis" }
 rustc_session = { path = "../rustc_session" }
 rustc_span = { path = "../rustc_span" }
 rustc_target = { path = "../rustc_target" }
-rustc_trait_selection = { path = "../rustc_trait_selection" }
 smallvec = { version = "1.8.1", features = ["union", "may_dangle"] }
 tracing = "0.1"
 # tidy-alphabetical-end
